Create valid JSON report
Open file with a square bracket to indicate a list.
Close the file with a square bracket to end the list.
Remove extra whitespace in each result.
Only one port is passed in with each result so do not make it
a list.
Remove the comma from the final item in the list.