Added a new feature that allows for file based locking. The purpose of this is to prevent multiple instances of masscan from running simultaneously (by accident) via an automated script. For example:
$ sudo masscan --range 10.0.0.0/8 \
               --ports 1-65535 \
               --rate 60000 \
               --lockfile /tmp/autoscan

This will have the effect of preventing another scan from running
when specifying the same lock file, assuming the initial scan is
still running / holding the lock.

This is useful when masscan is used with aggressive packet rates
on networks (LANs) that will melt if something goes awry and multiple
instances of the same scan kick off simultaneously, effectively
doubling (or tripling, etc.) the rate that packets are being sent.

This is a mitigation for edge cases in broken third party code, but
it seems like a simple enough feature to implement in the masscan
code itself.

The user has to obviously be aware of how to use the feature and
ensure that all invocations of masscan utilize the same lock file.

This is only tested on Linux, so it is only compiled in when the
__linux__ value is defined.