Fixed the boot status output to have correct information when an ICMP only scan is starting.
When performing a SYN scan, SYN + UDP scan, or SYN + UDP + ICMP scan,
the original behavior stays the same and reports a SYN Stealth scan.

This is just a cosmetic change to make it clearer that you didn't
invoke the scan incorrectly when trying to only perform a ping sweep.