grammar, spelling, and punctuation check
The first time reading through this my mind was blown...  🤯 I had to contribute somehow, so here's my first meager contribution... I hope nothing seems pedantic.  Let me know if you'd like explanations for any changes.  

One change which I anticipate might be unclear is the change from 'command-line' to 'command line', and leaving it in one place as "command-line", so I'll mention my rationale for that here.  When used as an adjective, the hyphen ties the two words together to form a single adjective, such as here:

> This is an implementation of the Nmap -oG output that can be easily parsed by command-line tools.

In the sentence above, "tools" is the noun and "command-line" tells us what kind of tools we're talking about.  If English used parentheses like Lisp (without the RPN ordering) we'd have something like `((command-line) tools)`.  But when using "command line" as a noun phrase (a noun and its adjective) the hyphen is not desirable, since adjectives do not normally get connected to their noun with a hyphen.